refactor(247): update routes code and architecture for tests
