+++ src/features/student/index.ts
@@ -1,2 +1,2 @@
 export * from './components'
-export { default as routes } from './routes'
+export * from './routes'
