+++ src/features/teacher/index.ts
@@ -1,1 +1,1 @@
-export { default as routes, TEACHER_HOME_ROUTE } from './routes'
+export * from './routes'
