+++ src/features/teacher/routes/index.ts
@@ -1,0 +1,2 @@
+export { default as routes } from './routes'
+export * from './routes'
